server/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-dotenv==1.1.0
supabase==2.15.2
python-multipart==0.0.20
numpy>=1.24.0
opencv-python-headless>=4.8.0
onnxruntime-gpu>=1.18.0
Pillow>=10.0.0
pydantic>=2.0.0
